feat: result aggregation part 4 - integration (#41)
When a non-streaming client sends a Message which initiates a long running task, the result is aggregated in-memory until a terminal event is received. Aggregation result is returned as a response.

Clients disconnects should not affect task execution. A cancel can be called while execution is in progress in which case it becomes a best effort operation.