feat: implement an a2aclient.Factory (#50)
Implement the logic for creating an a2aclient.Client configured with a compatible a2aclient.Transport.

Server preferences are defined as AgentCard.PreferredTransport first, then AgentCard.AdditionalInterfaces from first to last.
Client preferences can be specified through Config.PreferredTransports and will take precedence over server preferences.

To create a Client we find an intersection of client and server supported protocols and order them based on the client and server preferences.
We then try to establish a connection to one of the interfaces and fail if all attempts fail.