fix: Execute() callers missing events (#74)
### Details

Noticed that CI became flaky. Fails were easily reproducible with
`-count 1000`.
The cause was consumers missing events: 
1. Execution started.
2. Subscribe() called.
3. First events handled without any consumers to notify.
4. Subsribed.

Fixed by making `taskexec.Manager` create a default subscription and
return it from `Execute`.
The execution won't progress until this subscription is either canceled
or starts consuming events.

Additionally, made a couple of fixes to subscriptions:
1. Removed context from `subscription.cancel()`. If the method stops
trying to unsubscribe because context is canceled execution will get
stuck trying to write to the channel of this subscription.
2. Added reading from events channel during `subscription.cancel()`.
This handles the case when event broadcaster can't unregister the
subscription because it is trying to notify it about an event.
3. Added event broadcaster early error return for the case when errgroup
context gets canceled.
4. Changed subscription API such that `Events()` can only be consumed
once. Subscription automatically unregisters when consumption stops. I
think this way it should be harder to misuse.

I also update nightly workflow to run tests with`-count 100` and
parallelized `SendMessage` tests.