feat: task event factory (#95)
* Introduce `a2a.TaskInfoProvider` interface to simplify factory methods
usage by `AgentExecutor` implementers.
* `a2asrv.RequestContext` implements `a2a.TaskInfoProvider` so that a
`Task` or `TaskUpdateEvent` can be created from it.
* Rename `RequestContext.Task` to `RequestContext.StoredTask` to make it
more obvious that the field is unset for a fresh interaction. Before the
change it was easy to mistakenly start using `reqCtx.Task` as event
factory method argument.
* Introduce a `a2a.NewSubmittedTask` factory for creating a `Task` from
Message. It is easy to forget to store the message in history without
this utility.

The API required to create a Task for using event factory methods which
was bad because of the implicit knowledge about which fields need to be
initialized, for example:
```go
task := reqCtx.Task
if task == nil {
   task = &a2a.Task{ID: reqCtx.TaskID, ContextID: reqCtx.ContextID}
   event := a2a.NewStatusUpdateEvent(task, a2a.TaskStateSubmitted, nil)
   if err := queue.Write(ctx, event); err != nil {
      return fmt.Errorf("failed to setup a task: %w", err)
   }
}
```
with the change the same code looks like:
```go
if reqCtx.StoredTask == nil {
  event := a2a.NewStatusUpdateEvent(reqCtx, a2a.TaskStateSubmitted, nil)
  if err := queue.Write(ctx, event); err != nil {
    return fmt.Errorf("failed to setup a task: %w", err)
  }
}
```
We also don't need to carry this `&a2a.Task{ID: reqCtx.TaskID,
ContextID: reqCtx.ContextID}` task around for creating other events and
can just pass `reqCtx` as `TaskInfoProvider`.

re #84