feat: blocking flag handling (#97)
* Add `Polling` option to `a2aclient.Config`. If set to true `Blocking:
false` message configuration is sent to the server.
* Add `Blocking: false` handling on `a2asrv`. When execution results in
a Task, the Task is returned to the caller immediately.
* Make `Blocking` optional to match the "default to true if not set"
behavior all other SDKs implement.

Considered returning immediately when a requests references an existing
Task, but decided that'd be confusing to always receive the task in
`input-required` immediately after providing a follow-up.

re #96