refactor: event queue semantics (#120)
This is the second refactoring towards multi-process task processor (re
#115).

In the next PR we will decouple `func (e *localExecution) processEvents`
from `localExecution` to be able to reuse it with remote executions.

1. Introduce `eventpipe.Local`. This queue will only have a single
writer (AgentExecutor) and a single reader (event processor).
[commit](https://github.com/a2aproject/a2a-go/commit/a62a3788611224cf09e5fe77e9329804117c5e12)
2. Rewrite `eventqueue.Queue` to serve as an update broadcast mechanism.
Events are only written to it after they were successfully processed and
applied to the TaskStore. The in-memory implementation encapsulates
subscriber channel map previously managed by `localExecution`.
[commit](https://github.com/a2aproject/a2a-go/commit/ae71ee6be545eea97d50c3f978c0f6df39b55477)
3. Simplify `localSubscription` and `localExecution`. The logic
contained there is now encapsulated in `eventqueue.Queue`. Queues now
have a well-defined lifecycle. A queue is created before execution start
and destroyed after the execution end. When subscriptions are created
they use `Get` instead of `GetOrCreate` to tap the event stream.
[commit](https://github.com/a2aproject/a2a-go/commit/e5c41577920435cc8a00f6ce2f163d13878058ee)

This is potentially a breaking change if people were relying on
`eventqueue` package behavior outside of `a2asrv` context. But I doubt
this was the case.